+++ CHANGES.rst
@@ -15,6 +15,7 @@ upstream changes.
 
 Unreleased
 ----------
+ * **[FEATURE]** Added methods to sticky and unsticky posts 
  * **[FEATURE]** Add arguments syntax and time to
    :meth:`~praw.__init__.UnauthenticatedReddit.search`.
  * **[FEATURE]** PRAW will now try to use the http_proxy environment variable